Clarify "Additional configuration of blocks"
I, along with others, have gotten mixed up about the block config line. For a while I thought it was clarifying that I/O side config was what it was referring to, due to my eyes glancing over the "e.g." My request is to simply extend that "e.g." into two words to emphasize that things such as upgrades are also counted.