host servers
to be able to create severs easier without having to use community services, but then you ask me, how to create these services in a simple way. use creative editing as the basis. poker is very complicated you create a server by google. so it will be much easier for you to play with your unknown friend or person, leaving the servers open or private. but you can punish players like those in the communities.